server/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
python-dotenv==0.21.0
Scrapy==2.6.3
coverage==6.5.0
pytest==7.2.0
Psycopg2==2.9.5